+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/guide/SlimefunGuide.java
@@ -3,14 +3,14 @@ package io.github.thebusybiscuit.slimefun4.core.guide;
 import java.util.Optional;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
 import io.github.thebusybiscuit.slimefun4.api.player.PlayerProfile;
 import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
-import io.github.thebusybiscuit.slimefun4.implementation.guide.BookSlimefunGuide;
-import io.github.thebusybiscuit.slimefun4.implementation.guide.ChestSlimefunGuide;
+import io.github.thebusybiscuit.slimefun4.implementation.guide.SurvivalSlimefunGuide;
 import io.github.thebusybiscuit.slimefun4.utils.SlimefunUtils;
 import me.mrCookieSlime.Slimefun.Objects.Category;
 import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
@@ -22,36 +22,35 @@ import me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem;
  * @author TheBusyBiscuit
  * 
  * @see SlimefunGuideImplementation
- * @see ChestSlimefunGuide
- * @see BookSlimefunGuide
+ * @see SurvivalSlimefunGuide
  *
  */
 public final class SlimefunGuide {
 
     private SlimefunGuide() {}
 
+    @Nonnull
     public static ItemStack getItem(@Nonnull SlimefunGuideLayout design) {
         return SlimefunPlugin.getRegistry().getGuideLayout(design).getItem();
     }
 
-    public static void openCheatMenu(Player p) {
-        openMainMenuAsync(p, SlimefunGuideLayout.CHEAT_SHEET, 1);
+    public static void openCheatMenu(@Nonnull Player p) {
+        openMainMenuAsync(p, SlimefunGuideLayout.CHEAT_MODE, 1);
     }
 
-    public static void openGuide(Player p, ItemStack guide) {
-        if (SlimefunUtils.isItemSimilar(guide, getItem(SlimefunGuideLayout.CHEST), true)) {
-            openGuide(p, SlimefunGuideLayout.CHEST);
-        } else if (SlimefunUtils.isItemSimilar(guide, getItem(SlimefunGuideLayout.BOOK), true)) {
-            openGuide(p, SlimefunGuideLayout.BOOK);
-        } else if (SlimefunUtils.isItemSimilar(guide, getItem(SlimefunGuideLayout.CHEAT_SHEET), true)) {
-            openGuide(p, SlimefunGuideLayout.CHEAT_SHEET);
+    public static void openGuide(@Nonnull Player p, @Nullable ItemStack guide) {
+        if (SlimefunUtils.isItemSimilar(guide, getItem(SlimefunGuideLayout.CHEAT_MODE), true)) {
+            openGuide(p, SlimefunGuideLayout.CHEAT_MODE);
         } else {
-            // When using /sf cheat or /sf open_guide, ItemStack is null.
-            openGuide(p, SlimefunGuideLayout.CHEST);
+            /*
+             * When using /sf cheat or /sf open_guide the ItemStack is null anyway,
+             * so we don't even need to check here at this point.
+             */
+            openGuide(p, SlimefunGuideLayout.SURVIVAL_MODE);
         }
     }
 
-    public static void openGuide(Player p, SlimefunGuideLayout layout) {
+    public static void openGuide(@Nonnull Player p, @Nonnull SlimefunGuideLayout layout) {
         if (!SlimefunPlugin.getWorldSettingsService().isWorldEnabled(p.getWorld())) {
             return;
         }
@@ -86,32 +85,29 @@ public final class SlimefunGuide {
     }
 
     public static void openSearch(PlayerProfile profile, String input, boolean survival, boolean addToHistory) {
-        SlimefunGuideImplementation layout = SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.CHEST);
+        SlimefunGuideImplementation layout = SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.SURVIVAL_MODE);
 
         if (!survival) {
-            layout = SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.CHEAT_SHEET);
+            layout = SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.CHEAT_MODE);
         }
 
         layout.openSearch(profile, input, addToHistory);
     }
 
     public static void displayItem(PlayerProfile profile, ItemStack item, boolean addToHistory) {
-        SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.CHEST).displayItem(profile, item, 0, addToHistory);
+        SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.SURVIVAL_MODE).displayItem(profile, item, 0, addToHistory);
     }
 
     public static void displayItem(PlayerProfile profile, SlimefunItem item, boolean addToHistory) {
-        SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.CHEST).displayItem(profile, item, addToHistory);
+        SlimefunPlugin.getRegistry().getGuideLayout(SlimefunGuideLayout.SURVIVAL_MODE).displayItem(profile, item, addToHistory);
     }
 
-    public static boolean isGuideItem(ItemStack item) {
-        return SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideLayout.CHEST), true) || SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideLayout.BOOK), true) || SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideLayout.CHEAT_SHEET), true);
+    public static boolean isGuideItem(@Nonnull ItemStack item) {
+        return SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideLayout.SURVIVAL_MODE), true) || SlimefunUtils.isItemSimilar(item, getItem(SlimefunGuideLayout.CHEAT_MODE), true);
     }
 
+    @Nonnull
     public static SlimefunGuideLayout getDefaultLayout() {
-        if (SlimefunPlugin.getCfg().getBoolean("guide.default-view-book")) {
-            return SlimefunGuideLayout.BOOK;
-        } else {
-            return SlimefunGuideLayout.CHEST;
-        }
+        return SlimefunGuideLayout.SURVIVAL_MODE;
     }
 }