+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/guide/RecipeChoiceTask.java
@@ -21,7 +21,7 @@ import io.github.thebusybiscuit.slimefun4.implementation.SlimefunPlugin;
  * A {@link RecipeChoiceTask} is an asynchronously repeating task that cycles
  * through the different variants of {@link Material} that a {@link MaterialChoice} or {@link Tag} can represent.
  * 
- * It is used in the {@link ChestSlimefunGuide} for any {@link ItemStack} from Minecraft
+ * It is used in the {@link SurvivalSlimefunGuide} for any {@link ItemStack} from Minecraft
  * that accepts more than one {@link Material} in its {@link Recipe}.
  * 
  * @author TheBusyBiscuit