+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/listeners/SlimefunGuideListener.java
@@ -12,8 +12,8 @@ import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.inventory.ItemStack;
 
-import io.github.thebusybiscuit.slimefun4.api.events.SlimefunGuideOpenEvent;
 import io.github.thebusybiscuit.slimefun4.api.events.PlayerRightClickEvent;
+import io.github.thebusybiscuit.slimefun4.api.events.SlimefunGuideOpenEvent;
 import io.github.thebusybiscuit.slimefun4.core.guide.SlimefunGuide;
 import io.github.thebusybiscuit.slimefun4.core.guide.SlimefunGuideLayout;
 import io.github.thebusybiscuit.slimefun4.core.guide.options.SlimefunGuideSettings;
@@ -47,24 +47,20 @@ public class SlimefunGuideListener implements Listener {
     public void onInteract(PlayerRightClickEvent e) {
         Player p = e.getPlayer();
 
-        if (tryOpenGuide(p, e, SlimefunGuideLayout.BOOK) == Result.ALLOW) {
+        if (tryOpenGuide(p, e, SlimefunGuideLayout.SURVIVAL_MODE) == Result.ALLOW) {
             if (p.isSneaking()) {
                 SlimefunGuideSettings.openSettings(p, e.getItem());
             } else {
-                openGuide(p, e, SlimefunGuideLayout.BOOK);
+                openGuide(p, e, SlimefunGuideLayout.SURVIVAL_MODE);
             }
-        } else if (tryOpenGuide(p, e, SlimefunGuideLayout.CHEST) == Result.ALLOW) {
+        } else if (tryOpenGuide(p, e, SlimefunGuideLayout.CHEAT_MODE) == Result.ALLOW) {
             if (p.isSneaking()) {
                 SlimefunGuideSettings.openSettings(p, e.getItem());
             } else {
-                openGuide(p, e, SlimefunGuideLayout.CHEST);
-            }
-        } else if (tryOpenGuide(p, e, SlimefunGuideLayout.CHEAT_SHEET) == Result.ALLOW) {
-            if (p.isSneaking()) {
-                SlimefunGuideSettings.openSettings(p, e.getItem());
-            } else {
-                // We rather just run the command here,
-                // all necessary permission checks will be handled there.
+                /*
+                 * We rather just run the command here, all
+                 * necessary permission checks will be handled there.
+                 */
                 p.chat("/sf cheat");
             }
         }