+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 * Removed an old version of bStats
 * CraftBukkit is officially no longer supported, Slimefun will now be disabled on old builds of CraftBukkit
 * Removed the deprecated ItemManipulationAPI for BlockMenus
+* Removed the "Written Book" variant of the Slimefun Guide
 
 #### Fixes
 * Fixed a couple of compatibility issues with ItemsAdder